Реализована загрузка данных и добавление в базу: /universe/regions/{region_id}/ /universe/constellations/{constellation_id}/ /universe/systems/{system_id}/

+++ EVEOnline.Data/Models/Uconstellation.cs
@@ -9,11 +9,11 @@ public partial class Uconstellation
 
     public string Name { get; set; } = null!;
 
-    public long PositionX { get; set; }
+    public decimal PositionX { get; set; }
 
-    public long PositionY { get; set; }
+    public decimal PositionY { get; set; }
 
-    public long PositionZ { get; set; }
+    public decimal PositionZ { get; set; }
 
     public int RegionId { get; set; }
