+++ EVEOnline.Data/Models/Usystem.cs
@@ -9,11 +9,11 @@ public partial class Usystem
 
     public string Name { get; set; } = null!;
 
-    public long PositionX { get; set; }
+    public decimal PositionX { get; set; }
 
-    public long PositionY { get; set; }
+    public decimal PositionY { get; set; }
 
-    public long PositionZ { get; set; }
+    public decimal PositionZ { get; set; }
 
     public string? SecurityClass { get; set; }
